requirements.txt
Django~=5.0.4
djangorestframework~=3.15.1